backend/requirements.txt
# Backend requirements
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
alembic==1.13.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt==4.0.1
pydantic[email]==2.7.0
pydantic-settings==2.2.1
openai==1.3.7
python-multipart>=0.0.6
requests==2.31.0
httpx==0.25.2
pinecone-client>=2.2.2
sentence-transformers>=2.2.2
pandas>=2.0.0
scikit-learn>=1.2.0
numpy>=1.24.0
joblib>=1.2.0
email-validator>=2.0.0
